backend/requirements.txt
uuid
PyJWT
bcrypt
authlib
alembic
pymysql
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
python-multipart
pydantic-settings